Add summer term related object methods to models sws.Section.
