+++ restclients/tests.py
@@ -3,6 +3,7 @@ from django.utils import unittest
 from restclients.test.uwnetid.subscription import EmailForwardingTest
 from restclients.test.util.date_formator import FormatorTest
 from restclients.test.util.datetime_convertor import DatetimeConvertorTest
+from restclients.test.util.summer_term import SummerTermTest
 from restclients.test.hfs.idcard import HfsTest
 from restclients.test.library.mylibinfo import MyLibInfoTest
 from restclients.test.digitlib.curric import DigitLibTest
